magit-delete-directory: use copy of tramp-compat-delete-directory
Previously we determined the number of arguments `delete-directory'
accepts at compile time and defined this function accordingly.  Now
`magit-delete-directory' assume `delete-directory' supports RECURSIVE
and only if that causes `wrong-number-of-arguments' it does the
recursive delete itself.

This fixes #684.  The old code broke on 24.3.50 and also was more noisy
and ugly than what we have now.