travis: allow build on emacs-snapshot to fail
This is yet another kludge necessary to prevent travis from always
failing (without there actually being a problem with magit).

The failure is this:

  $ elpa-get mocker cl-lib
  ...
  Wrong type argument: arrayp, mocker

Actually that was the failure yesterday, currently building on
emacs-snapshot does not even produce any output or terminate.
This might be unrelated though.

This might be related to recent struct-ification of package.el
in Emacs-24.3.50.