Move 'unpulled' section down in status buffer
At the moment 'unpulled' sits between 'untracked files' and
'changes'. This makes working in a non-clean repository pretty awkward
if the list of unpulled changes is longer than one or two commits as
it splits the important information that a user needs when working on
an unclean repository: untracked files and changes.

This change moves 'unpulled commits' right above 'unpushed
commits'. That way both will keep their relative order, and 'untracked
files' and 'changes' are close together. For a clean repository the
order is not even changed.