move definition of magit-buffer-switch and magit-quit-window
`magit-quit-window' was defined in the wrong place because it was at
come point only used for branch manager.  Now that it is used for all
Magit buffers it does not make sense to keep it in the branch section.

`magit-buffer-switch' was defined in the wrong place from the very
start.