magit.el: change and add section headings
For every key-mode popup command there is now a corresponding section by
the same name.  These headings consist of one word ending with "-ing".

There are a few other sections ending with "-ing" (Rebasing, Resetting,
Diffing, and Grepping) but at least for some of these we might actually
want to create new key-mode popup commands.

I would like to eventually put all "-ing" sections under a common super
section but that is currently not possible because a few other sections
are mixed inbetween that aren't related to key-mode popups.

For some other sections it would also make sense for them to be nested
below another section.  Again it is not reasonable to do this now
because these sections are not next to each other.

A few of the new sections (ChangeLog, Extensions, Interactive Rebase,
and Interactive Resolve) contain but one or two functions.  In case of
ChangeLog the reason is that its function is missplaced.  In the other
cases this is because the were part of the Miscellaneous section which
also contained other things that should be defined in top-level
sections.