create magit-compat.el for Git and Emacs compatibility code
Move the compatibility code for older versions of GNU Emacs
and Git to the new library magit-compat.el.

For each function that is defined for older Emacsen note
when it was first added to Emacs.

I am aware that the `magit-configure-have-*' functions use
Yoda conditions and I am at peace with it.  ;-)