magit-delete-branch: restore old behaviour
When the user tries to delete the checked out branch and then replies
that she wants to checkout master first then do so *and* delete the
other branch.

Before the previous commit 2107722 that touched this code we already did
that and I believe it was not intentional to change that.