just one "Various Utilities" subsection
Move some trivial functions so that all functions in the "Utilities"
section that do not belong into one of the more specialized subsections
are in one place: the "Various Utilities" subsection.

Also see f49fd99eee294ed6e0e0308b8c69c7e605337d52
which sets the stage for this and other commits.