move magit-with-refresh into "Utilities">"Various Utilities"
The macro `magit-with-refresh' cannot be defined in the same place as
the "refresh" functions it makes use of because these definitions come
after `magit-with-refresh' is used for the first time.

While this macro is important and used a lot it is not significant to
deserve its own top-level section, so move it into the "Utilities"
section.

Its definition is now located later in the file than before but still
before its first use.  (I don't expect it to make sense to define a
function that uses this macro earlier than where it is now defined.)