Mark "Acting" and "Act" sections
Put the commands that directly map to some git porcelain subcommand
into their own dedicated sections (one per subcommand).

These sections are conceptionally subsections of the new "Acting"
top-level section, but because there are other unrelated sections
located in between the subcommand (sub)sections there two such sections
are actually required.  The headings "Acting (1)" and "Acting (2)" are
used for those sections.  Some of the 'subsections of section "Acting"'
are not neighbors of any other such sections and for that reason are
currently top-level sections.  The headings of these sections have
" (A)" as suffix to indicate that the are at least conceptionally part
of the "Acting" section.  E.g. "Diffing (A)".

For some git porcelain subcommands a "key-mode" popup exists.  For
others there is no such popup (yet, in some cases).  To differentiate
these two kinds of Magit->Git mappings the headings eigher have a "ing"
suffix (e.g. "Merging") if a popup exists.  Otherwise they don't have
that suffix (e.g. "Rebase").