update authorship information (on master)
The copyright notices have not been updated in a very long time.  By
now so many people have contributed that it is not feasible anymore
to have a copyright line in `magit.el' for everyone.  Now only current
and former maintainers are listed explicitly in this file.  In other
libraries some non-maintainers who made changes to a particular file
are also listed - including but not limited to the original author.

Right below the now shorter Copyright list is a new link to the list of
all contributors on github.

The header keyword "Original Author" is replaced by just "Author",
that's how it is commonly done.  All keywords no longer contain
whitespace, but use a dash instead.

Both Phil Jackson and Peter J Weisberg are now listed as former
maintainers.  Peter has been removed from the list of active maintainers
as he hasn't done much lately.  Nicolas Dudebout and I are added to the
list of active maintainers.  That list is sorted by activity.

Appropriate Copyright lines and Author keywords are added to files
besides `magit.el' (partly this is done in an earlier commit).

In `magit-svn.el' and `magit-blame.el' several people have already been
listed that have not made a single commit to those file.  I assume this
is because these files have been ported from other packages and therefor
did not remove anyone.

I have added myself as Author of `magit-wip.el' even though Ryan
C. Thompson wrote the first version, because his version was barely
usable.

If you are unhappy about a change please contact me.