add magit-assert-one-parent, replacing magit-choose-parent-id
As before, it is not possible to revert or cherry-pick a merge commit.
That is okay because these are advanced operations with potentially
surprising consequences.  We might support this in the future but it
is not a priority at all.

Unlike before, we don't have a function that does not do what its name
advertises (which used to result in dead code in magit-apply-commit).