magit-apply-commit: remove optional NOERASE argument
Previously when NOERASE was non-nil the output from the previous
`magit-run*' invocation was not removed from the *magit-process*
buffer when calling this function.

The only caller of this function which made use of that argument was
`magit-rewrite-finish-step' (which didn't always use it).

So allso remove `magit-rewrite-finish-step's FIRST-P argument.

Indead, in this case it would make more sense to preserve the previous
output a little longer than in others cases.  On the other hand this is
a kludge which cannot be locally contained.  To keep it would make the
cleanup that follows in the next commits much harder.

(Instead we should make it *always* possible to recover the output of
the last few git invocations.  I would be even better to present that
information in the status buffer, instead of having her resort to the
*magit-run* buffer.  I will address this one way or another eventually.)