remove with-magit-tmp-buffer
This macro had semantics very different from `with-temp-buffer', a
name that is to close to that of the latter, and no doc-string.

This could lead to tears, or at least confuse people.  Blinded by the
similarity of the names at least I did not initially see the semantic
difference.

There are multiple ways this could be addressed.  E.g. we could find a
more descriptive name and add a doc-string.  But since this macro is
only used twice it makes more sense to just remove the abstraction and
let the two callers do the necessary work themselves.  I think that in
this case _not_ moving the complexity out of sight actually improves
the readers understanding even though the code becomes a little
longer (especially considering that the abstraction would/was only
used once).

Yet another approach would be to improve `magit-run-git-async's
semantics.  This approach wasn't taken because I wanted to contain the
changes to a small area.