magit-highlight-trailing-whitespace: cache style itself
Instead of having `magit-mode` caching the indentation style (even when
it is not used) do the caching inside `magit-highlight-trailing-whitespace'.

Remove the local variable `magit-current-identation' that was previously
used for that purpose and instead use set the local value of of
`magit-highlight-trailing-whitespace'.  It is okay that the custom and
local value of that variable doesn't have the same form because the
local value is only ever used inside `magit-highlight-trailing-whitespace'.

Also use `when' instead of `if' since there is no ELSE.