Don't use an unqualified title for the unstaged changes
The title of the unstaged changes section was changed to be dynamic in
c0d6281d, but no explanation is given for the change.

Deviating from the standard git nomenclature should only be done when
there is a clear win for the user. In this specific case the change in
wording creates an unwanted ambiguity in the interface.

Fixes #727