revive credits
A member of the "Magit Owners Team" [1] has recently taken it upon
himself to "update" the authorship information, in effect *purging*
pretty much the entire list of contributors [2,3], stating that it
"[has] not been updated in a very long time".

This statement is patently false, since Phil Jackson updated it as
recently as February of this year [4].  Even if it were true, this isn't
the kind of decision that should be made unilaterally, without so much
as a courtesy email to the mailing list.

Removing others' copyright statements is not just a violation of the
GPL, it's also an disturbing display of disrespect for all the time and
effort our contributors have expended to improve Magit for the good of
us all.  Giving credit where it's due, besides being common courtesy,
is one of the cornerstones of FOSS development.

But [2,3] also states "it is not feasible anymore to have a copyright
line in `magit.el' for everyone", so... let's keep all the authorship
information in an AUTHORS file instead.

I've globally replaced the copyright headers with a copyright statement
crediting "The Magit Project Developers" (do note that this is *not*
equivalent to "copyright assignment") with the correct copyright year(s),
a pointer to the AUTHORS file, and an "Author:" line, of which the value
was procured using:

#+begin_src sh :results verbatim
  for i in *.el ; do
      printf "%s\n  %s\n\n" \
          "${i}" \
          "$(git log --pretty=format:'%h : [%ai] %aN <%aE>' -- "${i}" | tail -n 1)" ;
  done
#+end_src

An exception was made for "magit-wip.el", where Ryan C. Thompson may
have introduced the file, but Jonas Bernoulli did code pretty much
the whole thing.

Files containing code borrowed from other projects (specifically
"magit-blame.el" and "magit-compat.el") now state the relevant
projects' names, URLs and licenses, right below the "Author:" line.

The Makefile now contains targets which use git-log(1) to generate an
unabridged list of our contributors' names and email addresses, which
is combined with "AUTHORS.in" and written to "AUTHORS", which is always
included when rolling tarballs and installing.

To prevent incomplete/duplicate/outdated/erroneous authors' names and
email addresses, a ".mailmap" file was created, which is automatically
picked up by git-log(1) when using %aN/%aE instead of %an/%ae.

The only contributors for whom I couldn't find any conclusive info
(in a reasonable amount of time) are "rabio <rabiodev at o2.pl>" and
"acple <silentsphere110@gmail.com>", so these are included as-is.

And thus the only maintenance overhead left to keeping this list up to
date is running "make authors" and committing the updated AUTHORS file,
right before tagging a new release.

[1] https://github.com/magit?tab=members
[2] commit a9ed908addc2b2b1fcdd4aabefe6befd1d0505b9
[3] commit 59fd7f6d2edc1d63449c85f616e8d169313633ed
[4] commit ca0f48162e2d6eb6c447d9cc393c54cc43fadab5

Signed-off-by: Pieter Praet <pieter@praet.org>