`magit{,-annotated}-tag': don't use `magit-read-tag'
Existing tag names can't be reused, so there's really no need to list
them all (thereby parsing the output of git-tag(1)) as possible options
when tagging a rev.  Besides being perfectly sufficient, `read-string'
is much more performant in repos with large amounts of tags.

This is a partial revert of commit 7cff9f04.

Signed-off-by: Pieter Praet <pieter@praet.org>