`magit--branches-for-remote-repo': simplify
By moving the bulk of the work to git-branch(1), we can do away with
some ugly string matching.  Performance-wise, it's pretty much a no-op,
but then again, I haven't been able to test it with a remote which has a
ridiculous amount of branches.

Now if only the function were actually used anywhere...
(conceived in 5bfb2164, employed in 73f25ef4, retired in 2107722b)

Signed-off-by: Pieter Praet <pieter@praet.org>