magit.texi: remove "Customization" chapter
I've introduced this in 64c21ea5 (merged in cb7e0235), and since
then, it hasn't been touched, except for 87857670 and d2b16f21
which fixed some typos in the documentation but not in the actual
docstrings which it's based on.

Since (a correct version of) all this information is available in
the docstrings and easily discoverable using `customize', there's
no point in leaving it to languish in the docs.

Signed-off-by: Pieter Praet <pieter@praet.org>