`magit-log{,-long}': replace `nconc' with `append'
According to the commit message of baa09bb5, the author intended to
replace `nconc' with `append' in `magit-log-long', but the actual diff
does the inverse.

Based on the commit message and the author's comment @ bug 633 [1],
the intent was to replace `nconc' with `append' in `magit-log' (as
opposed to `magit-log-long').

So, partially revert commit baa09bb5 and make the intended change.

[1] https://github.com/magit/magit/issues/633#issuecomment-18315327

Signed-off-by: Pieter Praet <pieter@praet.org>