add magit-with-git-editor-setup
The respective functionality was previously part of
`magit-interactive-rebase'.  Currently this is still the only
function that makes use of this new macro - but that will change.

Add a doc-string to make the macro less confusing.  (That's another
reason why this macro is a good idea; it gives us a place to document
the behaviour.)