update modeline inside `magit-revert-buffer'
We have a mechanism for "updating buffers that need it": the refresh
mechanism implemented by the `magit-refresh-*' functions and macros.

Refreshing the modeline on behalf of vc did not make use of that but
instead relied on `magit-update-vc-modeline' being called explicitly.
Now it is refreshed whenever the regular refresh mechanism kicks in,
which can both positive or negative.

Positive because we don't have to identify each place where updating
the modeline would make sense; negative because it is a potentially
costly operation which is now done under circumstances that do not
necessarily call for it.

I will take care of the drawbacks by making this both less costly as
well as optional all together, in later commits.  (This commit already
makes refreshing slightly more efficient because it removes one looping
of the list of buffers.)

The name of this function now doesn't match all it does anymore.
That will also be taken care of.