add magit-refresh-file-buffer-hook
Call it in `magit-refresh-wrapper' for every buffer visiting a file in
the current buffer.

Bring back `magit-update-vc-modeline' as a wrapper around
`vc-find-file-hook' to update _a single_ buffer.  Do _not_ add it to the
hook's default value.  I don't expect many people to pay attention to
the Vc controlled part of the modeline (and if they do they will notice
and can add it themselves).  On the other hand many people complain
about bad performance and not using this function does improve that.
See issue #573.  We should implement our own "status in modeline" hook
function, see issue #240.

Add new function `magit-revert-buffer' to the hook.  It does the same
thing as the now removed `magit-revert-buffers' (plural) did until a
four commits ago - just for a single buffer.