create magit-log-edit.el from existing code
Create new library `magit-log-edit.el' from code that was previously
part of `magit.el'.  Due to how the involved libraries depend on one
another some functions have to be declared in `magit.el' for the time
being.

The moved code isn't modified, except that it is reordered and split
into sections.