merge toplevel sections "Acting (1)" and "Acting (2)"
Until recently the toplevel section "Log Edit Mode" was between these
two sections, but that was moved to a separate file.

To merge these two sections another toplevel that was between the two
had to be made a subsection.  In the process that section "ChangeLog"
was renamed to "Commit".  In the future it will also contain a command
`magit-commit', which will be a wrapper around `magit-log-edit' or the
respective function from `git-commit-mode'.