allow using magit-add-log during rebase
Like `magit-log-edit' this function detected an progress rebase and
offered to continue that instead of committing.  When the user reply
was "no" this function acted as if that means "no I want to do
anything at all".  `magit-log-edit' already understands it correctly
"no thanks, I really do want to commit".

Change this function to behave the same way.  Because this function
used `magit-log-edit' internally the test for an in-progress rebase
can be removed here.  There is some code before the call to
`magit-log-edit' but that only collects some information, has no side
effects, and is inexpensive.