refactor log commands
Prompt the user for the range in the `interactive' forms of
`magit-log-ranged' or `magit-log-long-ranged'.

Previously this was not only done outside of the `interactive' form
of these functions but inside of `magit-log'/`magit-log-long' which
are called by the former.

This is wrong for two reasons:

(1) Whenever possible minibuffer input should be read in the
    interactive form.

(2) `magit-log'/`magit-long-log' only ever used the range when
    called by `magit-log-ranged'/`magit-log-long-ranged'.

`magit-log'/`magit-log-long' now take a RANGE as first argument,
replacing the "ask the user for a range and use that" switch
previously used.

Since this change affects almost all lines seize the opportunity
and move the "Logging" section into the toplevel "Acting" section.