`magit-automatic-merge': remove
It's pretty much a duplicate of `magit-merge'.

The only differences are that it uses `magit-read-rev' +
`magit-guess-branch' instead of `magit-read-rev-with-default',
and it doesn't pass `magit-custom-options' to `magit-run-git'.

Also, it's no longer in use: Added in 7dce8dc8, removed in bdebdfcf,
re-added in 08ebe7df as a fix for issue #57, retired in 60a35bc2.

Signed-off-by: Pieter Praet <pieter@praet.org>