magit-with-git-editor-setup: add SERVER-WINDOW argument
This allows displaying a buffer created by 'emacsclient' to be displayed
differently when using this macro as opposed to the default configured
with the `server-window' option.

Also add an option `magit-server-window-for-rebase' which is used by
`magit-interactive-rebase' to set that SERVER-WINDOW.  In a later commit
`magit-server-window-for-commit' will also be added.

At least in the case of `*-for-commit' this is useful because it makes
sense to use e.g. `pop-to-buffer' here, even when one generally prefers
`switch-to-buffer-other-frame'.