Makefile: restore 'magit-pkg.el' target
And depend on `cl-lib'.

Pieter Praet pointed out I removed that target in afb15b75 even though
I intended to keep as it is needed for the Marmalade tarball.

Someone else (github is currently down) pointed out that we need to
depend on `cl-lib'.  Therefore we also need that file for the Melpa
package.