`magit-tag': s/"-a"/"--annotate"/g
Enabling the "-a" switch in `magit-key-mode-popup-tagging' adds
"--annotate" to `magit-custom-options', however `magit-tag'
checked for the presence of "-a" to decide whether or not to use
`magit-with-git-editor-setup' + `magit-run-git-async'.

As a result, creating an annotated tag by setting the "-a" flag in
the tagging popup opened a new frame instead of using the existing
one, and blocked the parent Emacs process.

Signed-off-by: Pieter Praet <pieter@praet.org>