add `magit-run-git-gui-blame-interactively' and `magit-run-git-gui-blame' functions
Functions for invoking git's own Tcl/Tk-based GUI for annotation
(a.k.a. "git gui blame").  Some people like to use this interface
instead of or in addition to the various blame modes available
within emacs.

On some distributions, git-gui is packaged separately to git core, so
you need to ensure that you have the appropriate package installed.
On openSUSE, Ubuntu, and Fedora, the package is called 'git-gui'.

Not bound to anything by default.