Magit now requires GNU Emacs 23.2 or greater
The decision to officially the support for older versions of Emacs was
unanimously reached here: https://github.com/magit/magit/issues/708.

Actually the old claim in README.md that Magit works with GNU Emacs 22.1
was wrong.  Some features of Magit *might* have worked with that version
but a lot of things would not have.

Remove some old (and incomplete) backward compatibility definitions.

Since it is possible (while unlikely) that some users have successfully
used Magit with Emacs 22 in the past and because we are removing code
that might previously have allowed using a subset of Magit add a version
check to magit.el.

Check that the minimal version requirement is satisfied; else raise an
error.  It is better to raise an error at this point than to have late
adapters figure out on their own why Magit stopped working with their
beloved ancient Emacs.  I would also rather have people complain about
this decision than having them file bug reports about things that
stopped working - in which case we might fail to notice right away that
the breakage is causes by the use of an older Emacs.

This check will also cause an error when XEmacs is used (as XEmacs isn't
at 23.2 yet, and likely will never be).  The same arguments apply here
than the ones made above regarding old versions of GNU Emacs.