magit-set-default-diff-options: suppress warning
Variable `magit-diff-options' is not defined yet when
`magit-set-default-diff-options' is first called because it is used to
set its initial value while it is being defined.  However the part of
the function that causes the warning is not used to do so.  By the time
that part is used the variable is also defined.