restore magit-add-log that was with magit-log-edit.el
But name it `magit-commit-add-log' and add a doc-string to make it more
obvious what this is all about.  The only other difference to the old
code is that the call to `magit-log-edit' is replaced with some code
that instead pops to the closest buffer whose mode is `git-commit-mode'
(and that an error is raised if that isn't found).