use background color to hightlight current item
Visualizing added and removed lines in diffs using the background color
conflicts with also using the background color to highlight the current
item.  This can be overcome by using overlays in both cases - but this
comes with a performance penalty, which might or might not be noticable.
Never-the-less this is now default.  The alternatives are listed in the
doc-string of the new `magit-diff-use-overlays' option.

`magit-wash-diffstats-postwork' was previously one of the places where
section headers got propertized.  The same headers are also propertized
elsewhere, so that functionality is remove.  We should investigate
whether this function and the `magit-diffstat-cached-sections' variable
are needed at all (anymore?).