`magit-run-git-gui-blame': s/magit-start-process/start-file-process/
The "minimal-emacs" branch removed a number of compatibility functions,
among which `magit-start-process', replacing it with `start-file-process'.

Between the time that branch was created (commit 83dac351) and merged
(commit b872cbf7), `magit-run-git-gui-blame' was added (commit 007da202),
which also makes use of `magit-start-process', and this wasn't corrected
when merging in the "minimal-emacs" branch.

Signed-off-by: Pieter Praet <pieter@praet.org>