`magit-create-branch': simplify
I added the `magit-create-branch-behaviour' option back in 2011
(commit 3cf57646), but made the horrible mistake of leaving the
original behavior ("create branch at HEAD") the default.
See e.g. this thread [1] from February 2013 on the ML.

Instead of making 'at-point the default, let's just get rid of
the option altogether and always try to find a rev at point first.

Note that I'm intentionally *not* using `magit-default-rev', as
that uses `magit-guess-branch', which is very likely to offer
the *previous* value of HEAD instead of the current one.

[1] mid:"87k3q9pm73.fsf@debian.org"

Signed-off-by: Pieter Praet <pieter@praet.org>