remove unnecessary uses of let in some functions
... in the "Diffs and Hunks" section.

Local bindings should only be used when they are necessary because e.g.
(a) the value is used more than once and is costly to calculate (because
of either the computation or visual noise) or (b) the value has to be
calculated before its use because doing retrieving the value and using
it in the same place would change the outcome to to due to scope
differences (e.g. a different buffer is current).

When let is used and neither one of these or other requirements call for
it one is left wondering whether that is indeed the case, or whether one
just overlooks some subtle detail.  This cost outweights the benefit of
implicitly documenting the purpose of the function being called with the
local variables as arguments.

This doesn't full apply here, but the implicit "argument documentation"
argument is very weak in cases like this:

  (let ((very-descriptive-variable (very-descriptive-function something)))
    (only-use-of-the very-descriptive-variable))